Unsuccesful Attempt 1
The code in Progress.md was an attempt at getting Stock.js to display the stock that the user inputs as the ticker. Unfortunately, this code did not work. I am currently trying to figure out how to get the StockSymbol to equal the user input.

The file named ticker.js has also been deleted.