User Generated Stock Input
This code adds the ability for the user to change the stock depicted on the chart based off of their input into the input box.

The two commits before this are pretty aweful attempts. I was making the simple mistake of not binding this to fetch stock in the return/onClick event. The broswer was expecting a function but instead receiving a string. Lesson learned. I know what that error message means now.