Style Changes
This commit changes the header/title on the plotly graph. The title now reads: 100 Day Chart (Daily Opens)

This title is more reflective of what the chart actually shows, which is just the opening price for each day.

I have also added a link to this github repo at the bottom of the screen, as well as acknowledgments to Alpha Vantage and Plotly